api/requirements.txt
argon2-cffi ~= 18.1.0
blinker ~= 1.4
boto3 ~= 1.5.12
flask-classful ~= 0.14.1
Flask ~= 0.12.2
Flask-Cors ~= 3.0.3
openstacksdk ~=0.9.18
passlib ~= 1.7.1
psycopg2 ~= 2.7.1
raven ~= 6.4.0
redis ~= 2.10.6
postgrespy ~= 0.3.1
pyyaml ~= 3.12
